refactor: move gate collector to internal/integration (test-only)
Per review: the gate collector is integration-test infrastructure, not a
production-importable library, so it does not belong in a generic
internal/interop package.

- Moved internal/interop -> internal/integration; package renamed
  interop -> integration.
- The pure gate (Evaluate, decodeMediaFrame, isAVCCKeyframe) and its unit
  tests stay in the default `go test` run (fast, relay-free feedback,
  including the B-frame timestamp invariant).
- The live Collector (which needs a relay) is now collect_test.go behind the
  `integration` build tag — a _test.go file, so production cannot import it.
  It is compiled only by `go test -tags=integration`.
- ci.yml: the integration job now also covers ./internal/integration/... so
  the Collector is compile-validated against the moqt API in CI.
