+++ .github/workflows/ci.yml
@@ -88,7 +88,7 @@ jobs:
       # real-QUIC stack has not been validated under the race detector yet —
       # revisit enabling it once confirmed clean.
       - name: Run integration tests
-        run: go test -tags=integration -count=1 -timeout 5m ./internal/relay/...
+        run: go test -tags=integration -count=1 -timeout 5m ./internal/relay/... ./internal/integration/...
 
   build:
     name: Build
